Added PageClasses.py - Stores all page layouts - Fixed main scan buttons not showing up after selecting home - Cleaned up code
